requirements.txt
chromadb
langchain
sentence_transformers
pypdf
textract
openai
tiktoken
faiss-cpu